docs: added examples and update READMED
